+++ src/Our.Umbraco.SearchSpellCheck/Constants.cs
@@ -13,6 +13,8 @@
             internal const string ConfigurationSection = "SearchSpellCheck";
             internal const string IndexName = "Our.Umbraco.SearchSpellCheck.IndexName";
             internal const string IndexedFields = "Our.Umbraco.SearchSpellCheck.IndexedFields";
+            internal const string BuildOnStartup = "Our.Umbraco.SearchSpellCheck.BuildOnStartup";
+            internal const string RebuildOnPublish = "Our.Umbraco.SearchSpellCheck.RebuildOnPublish";
             internal const string AutoRebuildIndex = "Our.Umbraco.SearchSpellCheck.AutoRebuildIndex";
             internal const string AutoRebuildDelay = "Our.Umbraco.SearchSpellCheck.AutoRebuildDelay";
             internal const string AutoRebuildRepeat = "Our.Umbraco.SearchSpellCheck.AutoRebuildRepeat";