+++ src/Our.Umbraco.SearchSpellCheck/SpellCheckOptions.cs
@@ -6,7 +6,9 @@ namespace Our.Umbraco.SearchSpellCheck
     {
         public string IndexName { get; set; } = Constants.Configuration.DefaultIndexName;
         public List<string> IndexedFields { get; set; } = new List<string>(new string[] { "nodeName" });
-        public bool AutoRebuildIndex { get; set; } = true;
+        public bool BuildOnStartup { get; set; } = true;
+        public bool RebuildOnPublish { get; set; } = true;
+        public bool AutoRebuildIndex { get; set; } = false;
         public int AutoRebuildDelay { get; set; } = 5;
         public int AutoRebuildRepeat { get; set; } = 30;
     }