+++ src/Our.Umbraco.SearchSpellCheck/Startup.cs
@@ -4,11 +4,14 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Our.Umbraco.SearchSpellCheck.Indexing;
 using Our.Umbraco.SearchSpellCheck.Interfaces;
+using Our.Umbraco.SearchSpellCheck.NotificationHandlers;
 using Our.Umbraco.SearchSpellCheck.RecurringTasks;
 using Our.Umbraco.SearchSpellCheck.Services;
 using Umbraco.Cms.Core.Composing;
 using Umbraco.Cms.Core.DependencyInjection;
+using Umbraco.Cms.Core.Notifications;
 using Umbraco.Cms.Infrastructure.Examine;
+using Umbraco.Extensions;
 
 namespace Our.Umbraco.SearchSpellCheck
 {
@@ -24,10 +27,23 @@ namespace Our.Umbraco.SearchSpellCheck
             builder.Services.AddExamineLuceneIndex<SpellCheckIndex, ConfigurationEnabledDirectoryFactory>(options.IndexName);
             builder.Services.AddSingleton<SpellCheckValueSetBuilder>();
             builder.Services.AddSingleton<IIndexPopulator, SpellCheckIndexPopulator>();
+            builder.Services.AddSingleton<IIndexRebuilder, SpellCheckIndexRebuilder>();
+            builder.Services.AddUnique<SpellCheckIndexRebuilder>();
 
             if (options.AutoRebuildIndex)
             {
                 builder.Services.AddHostedService<SpellCheckIndexRebuild>();
+            }
+
+            // Notification handlers
+            if (options.BuildOnStartup)
+            {
+                builder.AddNotificationHandler<UmbracoRequestBeginNotification, BuildIndexOnStartupHandler>();
+            }
+
+            if (options.RebuildOnPublish)
+            {
+                builder.AddNotificationHandler<ContentPublishedNotification, BuildIndexOnPublishedHandler>();
             }
 
             // Services